// src/lib/types/addProductTypes.ts
export type T_FormVariables = {
  product_name: string;
  price: number | string;
  product_type: string;
  discount_percentage: number | string;
  product_category: string;
  primary_image: string;
  secondry_image_1: string;
  secondry_image_2: string;
  secondry_image_3: string;
  product_description: string;
};